pygamer: move buttons to its own module
